TodoList logic

+++ TodoItems/Domain/Interfaces/ITodoList.cs
@@ -5,7 +5,7 @@
         void AddItem(int id, string title, string description, string category);
         void UpdateItem(int id, string description);
         void RemoveItem(int id);
-        void RegisterProgression(int id, DateTime dateTime, decimal percent);
+        void RegisterProgression(int id, DateTime dateTime, float percent);
         void PrintItems();
     }
 }
